requirements.txt
rpaframework==12.2.0
